+++ OAuthAPIKit/Classes/Core/Promise+DecodeTo.swift
@@ -7,18 +7,7 @@
 
 import PromiseKit
 import Moya
-import Mapper
 import SwiftyJSON
-
-extension Promise where T == Response {
-
-    internal func decode<ResponseType: Mappable>(to type: ResponseType.Type) -> Promise<ResponseType> {
-        return then(on: decodingQueue, { response -> Promise<ResponseType> in
-            return .value(try response.decode(to: type))
-        })
-    }
-
-}
 
 extension Promise where T == Response {
