+++ OAuthAPIKit/Classes/RxSwiftExtensions/Single+DecodeTo.swift
@@ -5,7 +5,6 @@
 //  Created by David on 2019/11/22.
 //
 
-import Mapper
 import Moya
 import RxSwift
 import SwiftyJSON
@@ -16,13 +15,6 @@ extension PrimitiveSequence where Trait == SingleTrait, Element == Response {
         return observeOn(decodingScheduler)
         .flatMap({ response -> Single<JSON> in
             return .just(try response.decodeToJSON())
-        })
-    }
-
-    internal func decode<T: Mappable>(to type: T.Type) -> Single<T> {
-        return observeOn(decodingScheduler)
-        .flatMap({ response -> Single<T> in
-            return .just(try response.decode(to: type))
         })
     }
